+++ .eslintrc.js
@@ -1,0 +1,43 @@
+module.exports = {
+    env: {
+        browser: true,
+        es6: true
+    },
+    extends: [
+      "eslint:recommended",
+      "plugin:react/recommended",
+      "plugin:@typescript-eslint/eslint-recommended",
+      "plugin:@typescript-eslint/recommended",
+      "plugin:import/errors",
+      "plugin:import/warnings",
+      "plugin:import/typescript"
+    ],
+    globals: {
+        Atomics: "readonly",
+        SharedArrayBuffer: "readonly"
+    },
+    parser: "@typescript-eslint/parser",
+    parserOptions: {
+        ecmaFeatures: {
+            jsx: true
+        },
+        ecmaVersion: 2018,
+        sourceType: "module",
+        project: "./tsconfig.json"
+    },
+    plugins: [
+        "@typescript-eslint",
+        "react"
+    ],
+    rules: {
+        "@typescript-eslint/indent": ["error", 2],
+        "@typescript-eslint/explicit-function-return-type": "off",
+        "@typescript-eslint/explicit-member-accessibility": "off",
+        "@typescript-eslint/no-empty-interface": "off"
+    },
+    settings: {
+        react: {
+            version: "detect" // Tells eslint-plugin-react to automatically detect the version of React to use
+        }
+    }
+};