+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# master
+
+- Use TypeScript to write main code.
+    - Now, we have type definitions.🎉
+- Update storybook from 3.3.12 to 5.1.9
+- Update dependent node version from 6.x to 12.x
+
 # v0.11.0
 
 - Support react-bootstrap v1.x.x (Bootstrap v4) 🎉